requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.3
gunicorn==20.0.0
psycopg2==2.8.4
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
whitenoise==4.1.4
